HFP-92 [fix] CodeRabbit Review
- Redis 역직렬화 타입 불일치로 인한 ClassCastException 수정
  - FreelancerEvaluationSummaryDto: Map<String, Double> → Map<String, Object> + safeNumber(Object) 헬퍼 추가 (Integer/Long/Double/null 안전 변환)
  - EmployerReviewSummaryResponseDto: 동일 패턴 적용, getOrDefault → safeNumber 변환, null Map 방어 추가
  - FreelancerProjectStatusStatsDto: Map<String, Integer> → Map<String, ?> + safeInt(Object) 헬퍼 추가 (Integer/Long/String/null 안전 변환)
  - EmployerProjectStatsResponseDto: 동일 패턴 적용
- 서비스 레이어 unchecked cast 타입 일치화
  - FreelancerReviewService, EmployerReviewService: cast → Map<String, Object>
  - FreelancerProjectService, EmployerProjectService: cast → Map<String, ?>
- 테스트 mock 데이터 타입 수정 (실제 Redis 동작 시뮬레이션)
  - 리뷰 관련 mock: Map<String, Double> → Map<String, Object>, Integer/Double 혼합 사용
  - 프로젝트 통계 mock: Map<String, Integer> → Map<String, Long>

+++ freebridge/user/src/test/java/com/fallguys/mypage/service/employer/EmployerProjectServiceTest.java
@@ -36,10 +36,10 @@ class EmployerProjectServiceTest {
         Long employerId = 1L;
         String redisKey = "employer:project:stats:" + employerId;
 
-        Map<String, Integer> mockStats = Map.of(
-            "totalProjects", 10,
-            "activeApplicants", 5,
-            "contractedFreelancers", 15
+        Map<String, Long> mockStats = Map.of(   // Redis/Jackson은 정수를 Long으로 역직렬화함
+            "totalProjects", 10L,
+            "activeApplicants", 5L,
+            "contractedFreelancers", 15L
         );
 
         when(redisTemplate.opsForValue()).thenReturn(valueOperations);
